backend/requirements.txt
aiomysql==0.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
fastapi-users==15.0.5
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.2.0
google-genai==1.73.1
googleapis-common-protos==1.74.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.13
Jinja2==3.1.6
lxml==6.1.0
makefun==1.16.0
MarkupSafe==3.0.3
oauthlib==3.3.1
proto-plus==1.27.2
protobuf==7.34.1
pwdlib==0.3.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
PyMySQL==1.1.2
pyparsing==3.3.2
python-docx==1.1.2
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
requests==2.33.1
requests-oauthlib==2.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.45.0
watchfiles==1.1.1
websockets==16.0
